Switching JsonCreator to leverage column names
Instead of column position, moving towards a dynamic mapping between
columns. This way we can rearrange the columns and not worry about
updating the script.

Also, I'm moving the logic of the data retrieval out of the google
lookup and instead opting for the Google Drive integration to return an
Array of rows.