Extracting constant for improved organization
I wanted to separate the concept of building an item from the concept
of an item. In this way, Locabulary::Items is responsible for building
an item. Whereas Locabulary::Item is the container for the
classification of those items.

This is a pattern that I found in [Avdi Grimm's `naught` gem][1]

[1]: https://github.com/avdi/naught "Avdi Grimm's `naught`"