Removing FacetedHierarchicalTreeSorter
This functionality was added when the specifications were to deal with
a nested hash of structs from blacklight. It turns out that the
rendering process for blacklight builds the nested hash. There is a
point earlier in the Rails code that has a flat array items instead.

I am not expunging this data via a rebase, but instead am keeping this
commit around in case we want to undo it and bring back the
functionality.