Extracting common logic for HierarchyProcessor
There was a duplication of logic in two different hierarchy processors.
By having a common processor, I hope to leverage consistent behavior.

It is a more complicated setup, but one that I believe helps as there
could be additional hierarchies that we must process.

This is also the bare minimum extraction effort to remove duplication
of logic. There are means of tidying up the interaction between
hierarchy processing and otherwise.