Favoring Dry::Schema over Dry::Validation
Oh Dry, I looked to adopt you early in your development cycle.  Already,
you were a great library.  But you changed so much during your pre 1.0
days, and left many a breaking change for us to pick up.

This commit attempts to leverage the more up to date and more stable
version (due to Dry's commitment to semantic versioning).

In addition, I've moved the schema for the JSON document into the specs
themselves.  In this way, upstream implementations don't need to worry
about the schema (which is relevant only to the JSON documents that we
store locally within this gem).

I needed to do some antics for 2.3.8 support and code coverage.